Move the `Ticket` model into the `Conference` app
