+++ backend/conferences/models/__init__.py
@@ -1,0 +1,6 @@
+from .audience_level import AudienceLevel
+from .conference import Conference
+from .deadline import Deadline
+from .duration import Duration
+from .topic import Topic
+from .ticket import Ticket
