+++ backend/conferences/models/ticket.py
@@ -8,10 +8,10 @@ from model_utils.models import TimeFramedModel, TimeStampedModel
 class Ticket(TimeFramedModel, TimeStampedModel):
     conference = models.ForeignKey('conferences.Conference', on_delete=models.CASCADE, verbose_name=_('conference'), related_name='tickets')
 
-    name = models.CharField(_('name'), max_length=100)  # Early Bird Student
+    code = models.CharField(_('code'), max_length=10)
+    name = models.CharField(_('name'), max_length=100)
     description = models.TextField(_('description'))
     price = models.DecimalField(_('price'), max_digits=10, decimal_places=2)
-    code = models.CharField(_('code'), max_length=10)
 
     def __str__(self):
         return f'{self.name} ({self.conference.name})'
