+++ backend/pycon/settings/base.py
@@ -30,7 +30,6 @@ INSTALLED_APPS = [
 
     'users',
     'conferences.apps.ConferencesConfig',
-    'tickets.apps.TicketsConfig',
     'languages.apps.LanguagesConfig',
     'submissions.apps.SubmissionsConfig',
 ]
