+++ backend/tests/conftest.py
@@ -5,7 +5,6 @@ from languages.models import Language
 
 from .api.fixtures import *  # noqa
 from .conferences.factories import *  # noqa
-from .tickets.factories import *  # noqa
 from .submissions.factories import *  # noqa
