--- backend/tickets/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class TicketsConfig(AppConfig):
-    name = 'tickets'
